requirements.txt
openai
moviepy>=2.0.0
gtts
Pillow
google-api-python-client
google-auth
google-auth-oauthlib
imageio
imageio-ffmpeg
